moved sub somponents
